iiii

+++ src/d_net.c
@@ -77,7 +77,7 @@ unsigned NetbufferChecksum (void)
   return 0;                       /* byte order problems */
 #endif
   
-  l = (NetbufferSize () - (int)&(((doomdata_t *)0)->retransmitfrom))/4;
+  l = (int)((NetbufferSize () - (unsigned long)&(((doomdata_t *)0)->retransmitfrom))/4);
   for (i=0 ; i<l ; i++)
     c += ((unsigned *)&netbuffer->retransmitfrom)[i] * (i+1);
